// frontend-mobile/src/types/models.ts
export interface MapStop {
  lat: number;
  lon: number;
  patient_name: string;
  care_type: string;
  time: string;
  sector_name: string;
  sector_color: string;
}

export interface MapData {
  stops: MapStop[];
  center_lat: number;
  center_lon: number;
}